Move timings to build and improve tests
