+++ coderd/workspaces.go
@@ -1740,77 +1740,6 @@ func (api *API) watchWorkspace(rw http.ResponseWriter, r *http.Request) {
 	}
 }
 
-// @Summary Get workspace timings by ID
-// @ID get-workspace-timings-by-id
-// @Security CoderSessionToken
-// @Produce json
-// @Tags Workspaces
-// @Param workspace path string true "Workspace ID" format(uuid)
-// @Success 200 {object} codersdk.WorkspaceTimings
-// @Router /workspaces/{workspace}/timings [get]
-func (api *API) workspaceTimings(rw http.ResponseWriter, r *http.Request) {
-	var (
-		ctx       = r.Context()
-		workspace = httpmw.WorkspaceParam(r)
-	)
-
-	build, err := api.Database.GetLatestWorkspaceBuildByWorkspaceID(ctx, workspace.ID)
-	if err != nil {
-		httpapi.Write(ctx, rw, http.StatusInternalServerError, codersdk.Response{
-			Message: "Internal error fetching workspace build.",
-			Detail:  err.Error(),
-		})
-		return
-	}
-
-	provisionerTimings, err := api.Database.GetProvisionerJobTimingsByJobID(ctx, build.JobID)
-	if err != nil && !errors.Is(err, sql.ErrNoRows) {
-		httpapi.Write(ctx, rw, http.StatusInternalServerError, codersdk.Response{
-			Message: "Internal error fetching workspace timings.",
-			Detail:  err.Error(),
-		})
-		return
-	}
-
-	agentScriptTimings, err := api.Database.GetWorkspaceAgentScriptTimingsByBuildID(ctx, build.ID)
-	if err != nil && !errors.Is(err, sql.ErrNoRows) {
-		httpapi.Write(ctx, rw, http.StatusInternalServerError, codersdk.Response{
-			Message: "Internal error fetching workspace agent script timings.",
-			Detail:  err.Error(),
-		})
-		return
-	}
-
-	res := codersdk.WorkspaceTimings{
-		ProvisionerTimings: make([]codersdk.ProvisionerTiming, 0, len(provisionerTimings)),
-		AgentScriptTimings: make([]codersdk.AgentScriptTiming, 0, len(agentScriptTimings)),
-	}
-
-	for _, t := range provisionerTimings {
-		res.ProvisionerTimings = append(res.ProvisionerTimings, codersdk.ProvisionerTiming{
-			JobID:     t.JobID,
-			Stage:     string(t.Stage),
-			Source:    t.Source,
-			Action:    t.Action,
-			Resource:  t.Resource,
-			StartedAt: t.StartedAt,
-			EndedAt:   t.EndedAt,
-		})
-	}
-	for _, t := range agentScriptTimings {
-		res.AgentScriptTimings = append(res.AgentScriptTimings, codersdk.AgentScriptTiming{
-			StartedAt:   t.StartedAt,
-			EndedAt:     t.EndedAt,
-			ExitCode:    t.ExitCode,
-			Stage:       string(t.Stage),
-			Status:      string(t.Status),
-			DisplayName: t.DisplayName,
-		})
-	}
-
-	httpapi.Write(ctx, rw, http.StatusOK, res)
-}
-
 type workspaceData struct {
 	templates    []database.Template
 	builds       []codersdk.WorkspaceBuild
