+++ coderd/workspaces_test.go
@@ -3556,240 +3556,3 @@ func TestWorkspaceNotifications(t *testing.T) {
 		})
 	})
 }
-
-func TestWorkspaceTimings(t *testing.T) {
-	t.Parallel()
-
-	// Setup a base template for the workspaces
-	db, pubsub := dbtestutil.NewDB(t)
-	client := coderdtest.New(t, &coderdtest.Options{
-		Database: db,
-		Pubsub:   pubsub,
-	})
-	owner := coderdtest.CreateFirstUser(t, client)
-	file := dbgen.File(t, db, database.File{
-		CreatedBy: owner.UserID,
-	})
-	versionJob := dbgen.ProvisionerJob(t, db, pubsub, database.ProvisionerJob{
-		OrganizationID: owner.OrganizationID,
-		InitiatorID:    owner.UserID,
-		WorkerID:       uuid.NullUUID{},
-		FileID:         file.ID,
-		Tags: database.StringMap{
-			"custom": "true",
-		},
-	})
-	version := dbgen.TemplateVersion(t, db, database.TemplateVersion{
-		OrganizationID: owner.OrganizationID,
-		JobID:          versionJob.ID,
-		CreatedBy:      owner.UserID,
-	})
-	template := dbgen.Template(t, db, database.Template{
-		OrganizationID:  owner.OrganizationID,
-		ActiveVersionID: version.ID,
-		CreatedBy:       owner.UserID,
-	})
-
-	// Since the tests run in parallel, we need to create a new workspace for
-	// each test to avoid fetching the wrong latest build.
-	type testWorkspace struct {
-		database.Workspace
-		build  database.WorkspaceBuild
-		script database.WorkspaceAgentScript
-	}
-	makeWorkspace := func() testWorkspace {
-		ws := dbgen.Workspace(t, db, database.Workspace{
-			OwnerID:        owner.UserID,
-			OrganizationID: owner.OrganizationID,
-			TemplateID:     template.ID,
-			// Generate unique name for the workspace
-			Name: "test-workspace-" + uuid.New().String(),
-		})
-		jobID := uuid.New()
-		job := dbgen.ProvisionerJob(t, db, pubsub, database.ProvisionerJob{
-			ID:             jobID,
-			OrganizationID: owner.OrganizationID,
-			Type:           database.ProvisionerJobTypeWorkspaceBuild,
-			Tags:           database.StringMap{jobID.String(): "true"},
-		})
-		build := dbgen.WorkspaceBuild(t, db, database.WorkspaceBuild{
-			WorkspaceID:       ws.ID,
-			TemplateVersionID: version.ID,
-			BuildNumber:       1,
-			Transition:        database.WorkspaceTransitionStart,
-			InitiatorID:       owner.UserID,
-			JobID:             job.ID,
-		})
-		// Create a resource, agent, and script to test the timing of agent scripts
-		resource := dbgen.WorkspaceResource(t, db, database.WorkspaceResource{
-			JobID: jobID,
-		})
-		agent := dbgen.WorkspaceAgent(t, db, database.WorkspaceAgent{
-			ResourceID: resource.ID,
-		})
-		scripts := dbgen.WorkspaceAgentScripts(t, db, database.InsertWorkspaceAgentScriptsParams{
-			WorkspaceAgentID: agent.ID,
-			CreatedAt:        time.Now(),
-			LogSourceID: []uuid.UUID{
-				uuid.New(),
-			},
-			LogPath:          []string{""},
-			Script:           []string{""},
-			Cron:             []string{""},
-			StartBlocksLogin: []bool{false},
-			RunOnStart:       []bool{false},
-			RunOnStop:        []bool{false},
-			TimeoutSeconds:   []int32{0},
-			DisplayName:      []string{""},
-			ID: []uuid.UUID{
-				uuid.New(),
-			},
-		})
-
-		return testWorkspace{
-			Workspace: ws,
-			build:     build,
-			script:    scripts[0],
-		}
-	}
-
-	makeProvisionerTimings := func(jobID uuid.UUID, count int) []database.ProvisionerJobTiming {
-		// Use the database.ProvisionerJobTiming struct to mock timings data instead
-		// of directly creating database.InsertProvisionerJobTimingsParams. This
-		// approach makes the mock data easier to understand, as
-		// database.InsertProvisionerJobTimingsParams requires slices of each field
-		// for batch inserts.
-		timings := make([]database.ProvisionerJobTiming, count)
-		now := time.Now()
-		for i := range count {
-			startedAt := now.Add(-time.Hour + time.Duration(i)*time.Minute)
-			endedAt := startedAt.Add(time.Minute)
-			timings[i] = database.ProvisionerJobTiming{
-				StartedAt: startedAt,
-				EndedAt:   endedAt,
-				Stage:     database.ProvisionerJobTimingStageInit,
-				Action:    string(database.AuditActionCreate),
-				Source:    "source",
-				Resource:  fmt.Sprintf("resource[%d]", i),
-			}
-		}
-		insertParams := database.InsertProvisionerJobTimingsParams{
-			JobID: jobID,
-		}
-		for _, timing := range timings {
-			insertParams.StartedAt = append(insertParams.StartedAt, timing.StartedAt)
-			insertParams.EndedAt = append(insertParams.EndedAt, timing.EndedAt)
-			insertParams.Stage = append(insertParams.Stage, timing.Stage)
-			insertParams.Action = append(insertParams.Action, timing.Action)
-			insertParams.Source = append(insertParams.Source, timing.Source)
-			insertParams.Resource = append(insertParams.Resource, timing.Resource)
-		}
-		return dbgen.ProvisionerJobTimings(t, db, insertParams)
-	}
-
-	makeAgentScriptTimings := func(scriptID uuid.UUID, count int) []database.WorkspaceAgentScriptTiming {
-		newTimings := make([]database.InsertWorkspaceAgentScriptTimingsParams, count)
-		now := time.Now()
-		for i := range count {
-			startedAt := now.Add(-time.Hour + time.Duration(i)*time.Minute)
-			endedAt := startedAt.Add(time.Minute)
-			newTimings[i] = database.InsertWorkspaceAgentScriptTimingsParams{
-				StartedAt: startedAt,
-				EndedAt:   endedAt,
-				Stage:     database.WorkspaceAgentScriptTimingStageStart,
-				ScriptID:  scriptID,
-				ExitCode:  0,
-				Status:    database.WorkspaceAgentScriptTimingStatusOk,
-			}
-		}
-
-		timings := make([]database.WorkspaceAgentScriptTiming, 0)
-		for _, newTiming := range newTimings {
-			timing := dbgen.WorkspaceAgentScriptTiming(t, db, newTiming)
-			timings = append(timings, timing)
-		}
-
-		return timings
-	}
-
-	// Given
-	testCases := []struct {
-		provisionerTimings int
-		agentScriptTimings int
-		workspace          testWorkspace
-	}{
-		{
-			workspace: makeWorkspace(),
-		},
-		{
-			provisionerTimings: 5,
-			workspace:          makeWorkspace(),
-		},
-		{
-			provisionerTimings: 2,
-			workspace:          makeWorkspace(),
-		},
-		{
-			agentScriptTimings: 5,
-			workspace:          makeWorkspace(),
-		},
-		{
-			agentScriptTimings: 2,
-			workspace:          makeWorkspace(),
-		},
-		{
-			provisionerTimings: 3,
-			agentScriptTimings: 4,
-			workspace:          makeWorkspace(),
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-		name := fmt.Sprintf("provisionerTimings=%d, agentScriptTimings=%d", tc.provisionerTimings, tc.agentScriptTimings)
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			// Generate timings based on test config
-			genProvisionerTimings := makeProvisionerTimings(tc.workspace.build.JobID, tc.provisionerTimings)
-			genAgentScriptTimings := makeAgentScriptTimings(tc.workspace.script.ID, tc.agentScriptTimings)
-
-			res, err := client.WorkspaceTimings(context.Background(), tc.workspace.ID)
-			require.NoError(t, err)
-			require.Len(t, res.ProvisionerTimings, tc.provisionerTimings)
-			require.Len(t, res.AgentScriptTimings, tc.agentScriptTimings)
-
-			// Verify timings data
-			for i := range res.ProvisionerTimings {
-				timingRes := res.ProvisionerTimings[i]
-				genTiming := genProvisionerTimings[i]
-				require.Equal(t, genTiming.Resource, timingRes.Resource)
-				require.Equal(t, genTiming.Action, timingRes.Action)
-				require.Equal(t, string(genTiming.Stage), timingRes.Stage)
-				require.Equal(t, genTiming.JobID.String(), timingRes.JobID.String())
-				require.Equal(t, genTiming.Source, timingRes.Source)
-				require.Equal(t, genTiming.StartedAt.UnixMilli(), timingRes.StartedAt.UnixMilli())
-				require.Equal(t, genTiming.EndedAt.UnixMilli(), timingRes.EndedAt.UnixMilli())
-			}
-			for i := range res.AgentScriptTimings {
-				timingRes := res.AgentScriptTimings[i]
-				genTiming := genAgentScriptTimings[i]
-				require.Equal(t, genTiming.ExitCode, timingRes.ExitCode)
-				require.Equal(t, string(genTiming.Status), timingRes.Status)
-				require.Equal(t, string(genTiming.Stage), timingRes.Stage)
-				require.Equal(t, genTiming.StartedAt.UnixMilli(), timingRes.StartedAt.UnixMilli())
-				require.Equal(t, genTiming.EndedAt.UnixMilli(), timingRes.EndedAt.UnixMilli())
-			}
-		})
-	}
-}
-
-func TestWorkspaceTimings_NotFound(t *testing.T) {
-	t.Parallel()
-
-	client := coderdtest.New(t, nil)
-	coderdtest.CreateFirstUser(t, client)
-
-	_, err := client.WorkspaceTimings(context.Background(), uuid.New())
-	require.Contains(t, err.Error(), "not found")
-}
