+++ codersdk/workspaces.go
@@ -626,44 +626,6 @@ func (c *Client) UnfavoriteWorkspace(ctx context.Context, workspaceID uuid.UUID)
 	return nil
 }
 
-type ProvisionerTiming struct {
-	JobID     uuid.UUID `json:"job_id" format:"uuid"`
-	StartedAt time.Time `json:"started_at" format:"date-time"`
-	EndedAt   time.Time `json:"ended_at" format:"date-time"`
-	Stage     string    `json:"stage"`
-	Source    string    `json:"source"`
-	Action    string    `json:"action"`
-	Resource  string    `json:"resource"`
-}
-
-type AgentScriptTiming struct {
-	StartedAt   time.Time `json:"started_at"`
-	EndedAt     time.Time `json:"ended_at"`
-	ExitCode    int32     `json:"exit_code"`
-	Stage       string    `json:"stage"`
-	Status      string    `json:"status"`
-	DisplayName string    `json:"display_name"`
-}
-
-type WorkspaceTimings struct {
-	ProvisionerTimings []ProvisionerTiming `json:"provisioner_timings"`
-	AgentScriptTimings []AgentScriptTiming `json:"agent_script_timings"`
-}
-
-func (c *Client) WorkspaceTimings(ctx context.Context, id uuid.UUID) (WorkspaceTimings, error) {
-	path := fmt.Sprintf("/api/v2/workspaces/%s/timings", id.String())
-	res, err := c.Request(ctx, http.MethodGet, path, nil)
-	if err != nil {
-		return WorkspaceTimings{}, err
-	}
-	defer res.Body.Close()
-	if res.StatusCode != http.StatusOK {
-		return WorkspaceTimings{}, ReadBodyAsError(res)
-	}
-	var timings WorkspaceTimings
-	return timings, json.NewDecoder(res.Body).Decode(&timings)
-}
-
 // WorkspaceNotifyChannel is the PostgreSQL NOTIFY
 // channel to listen for updates on. The payload is empty,
 // because the size of a workspace payload can be very large.
